+++ src/view/__init__.py
@@ -16,8 +16,10 @@ from . import _codec
 from .__about__ import __license__, __version__
 from .app import *
 from .components import *
+from .databases import *
 from .exceptions import *
 from .logging import *
+from .patterns import *
 from .response import *
 from .routing import *
 from .util import *