+++ tests/simple_routes/third/fourth.py
@@ -1,0 +1,11 @@
+from view import delete, options
+
+
+@delete("/delete")
+async def d():
+    return "delete"
+
+
+@options("/options")
+async def o():
+    return "options"