+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added `get_app`
 - Added documentation generation
+- Added database support
+- Removed `attempt_import` and `MissingLibraryError`
 - Added support for lists in type validation
 - Added support for implicit query parameters
 - Renamed `debug` to `enable_debug`
@@ -17,6 +19,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed `EnvironmentError` to `BadEnvironmentError`
 - Added logging functions to `App`
 - Changed environment prefixes for configuration
+- Added `patterns` loader
+- Added handling of relative paths in the configuration setting `loader_path`
+- Added exists validation to `loader_path`
+- Add path to `PATH` environment variable during loading
 
 ## [1.0.0-alpha5] - 2023-09-24